+++ allennlp_models/classification/__init__.py
@@ -1,4 +1,3 @@
 # flake8: noqa: F403
 from allennlp_models.classification.models import *
 from allennlp_models.classification.dataset_readers import *
-from allennlp_models.classification.tango import *