+++ allennlp_models/classification/tango/__init__.py
@@ -1,1 +1,0 @@
-from allennlp_models.classification.tango.imdb import ImdbInstances