+++ allennlp_models/mc/tango/__init__.py
@@ -1,1 +1,0 @@
-from allennlp_models.mc.tango.piqa import PiqaInstances