+++ tests/training_config_test.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 from glob import glob
 import os
-from typing import Dict, Tuple
+from typing import Dict, Tuple, Set
 
 import pytest
 
@@ -18,10 +18,7 @@ CONFIGS_TO_IGNORE = {
     "constituency_parser_transformer_elmo.jsonnet",
 }
 
-FOLDERS_TO_IGNORE = {
-    # TODO (epwalsh/dirkg): need to test tango configs differently.
-    "tango",
-}
+FOLDERS_TO_IGNORE: Set[str] = set()
 
 
 def find_configs():