+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Following a breaking change in the NLTK API, we now depend on the most recent version only.
+- Uses Tango from https://github.com/allenai/tango instead of from AllenNLP
 
 ### Removed