+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v2.9.0](https://github.com/allenai/allennlp-models/releases/tag/v2.9.0) - 2022-01-27
+
 ### Added
 
 - Added Python 3.9 to the testing matrix